add VALID and INVALID states to bundler
