+++ packages/server/src/orchestrator/state.ts
@@ -1,5 +1,5 @@
 import { Test, TestResult, ResultStatus } from '@bigtest/suite';
-import { BundlerError, BundlerWarning } from '@bigtest/bundler';
+import { BundlerError, BundlerWarning, ValidationWarning, ValidationError } from '@bigtest/bundler';
 
 export type AgentState = {
   agentId: string;
@@ -37,45 +37,32 @@ export type TestRunAgentState = {
 export type BundlerState =
   | { type: 'UNBUNDLED' }
   | { type: 'BUILDING'; warnings: BundlerWarning[] }
+  | { type: 'VALIDATING' }
+  | { type: 'INVALID'; errors: ValidationError[]; warnings: ValidationWarning[] }
+  | { type: 'VALID'; warnings: ValidationWarning[] }
   | { type: 'GREEN'; path: string;  warnings: BundlerWarning[] }
   | { type: 'ERRORED'; error: BundlerError }
 
 export type BundlerTypes = Pick<BundlerState, 'type'>['type'];
 
-// TODO: All errors could implement at least these fields?
-export type ValidationException = {
-  fileName: string;
-  message: string;
-  displayMessage?: string;
-  code?: string;
-  frame?: string;
-	loc?: {
-    column: number;
-		file?: string;
-		line: number;
-  };
-}
-
-export type ValidationWarning = ValidationException;
-export type ValidationError = ValidationException & {
-  stack?: string;
-}
-
 export type ValidatorState = 
-| { type: 'IDLE' }
 | { type: 'VALIDATING' }
 | { type: 'VALID';  warnings: ValidationWarning[] }
 | { type: 'INVALID'; errors: ValidationError[]; warnings: ValidationWarning[] }
 
+export interface Validator {
+  validate(files: string[]): BundlerState;
+}
+
 export type ValidatorStates = Pick<ValidatorState, 'type'>['type'];
 
 export interface Validator {
-  validate(files: string[]): ValidatorState;
+  validate(files: string[]): BundlerState;
 }
 
 export interface Manifest extends Test {
   fileName: string;
-  validState: ValidatorState;
+  errors: ValidationError[];
 }
 
 export type AppStatus = 'unstarted' | 'started' | 'reachable' | 'unreachable' | 'crashed'
