+++ packages/server/test/command-server.test.ts
@@ -151,7 +151,7 @@ describe('command server', () => {
         steps: [],
         assertions: [],
         children: [test1, test2],
-        validState: { type: 'IDLE' },
+        errors: [],
       });
     });
