+++ packages/server/test/manifest-generator.test.ts
@@ -4,8 +4,6 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as rmrf from 'rimraf';
 
-import { Test } from '@bigtest/suite';
-
 import { actions } from './helpers';
 
 import { createManifestGenerator } from '../src/manifest-generator';
@@ -26,7 +24,7 @@ async function loadManifest() {
   return await import(fullPath);
 }
 
-describe('manifest-generator', () => {
+describe.only('manifest-generator', () => {
   let delegate: Mailbox;
   let atom: Atom<OrchestratorState>;
 
@@ -51,7 +49,7 @@ describe('manifest-generator', () => {
   });
 
   describe('starting', () => {
-    let manifest: Test;
+    let manifest: Manifest;
 
     beforeEach(async () => {
       manifest = await loadManifest();
@@ -61,11 +59,12 @@ describe('manifest-generator', () => {
       expect(manifest.children.length).toEqual(2)
       expect(manifest.children[0]).toEqual({ path: './tmp/manifest-generator/test1.t.js', description: 'hello' });
       expect(manifest.children[1]).toEqual({ path: './tmp/manifest-generator/test2.t.js', description: 'monkey' });
+      expect(manifest.errors).toHaveLength(0);
     });
   });
 
   describe('adding a test file', () => {
-    let manifest: Test;
+    let manifest: Manifest;
 
     beforeEach(async () => {
       await writeFile(join(TEST_DIR, "/test3.t.js"), "module.exports = { default: { description: 'test' } };");
@@ -78,11 +77,12 @@ describe('manifest-generator', () => {
       expect(manifest.children[0]).toEqual({ path: './tmp/manifest-generator/test1.t.js', description: 'hello' });
       expect(manifest.children[1]).toEqual({ path: './tmp/manifest-generator/test2.t.js', description: 'monkey' });
       expect(manifest.children[2]).toEqual({ path: './tmp/manifest-generator/test3.t.js', description: 'test' });
+      expect(manifest.errors).toHaveLength(0);
     });
   });
 
   describe('removing a test file', () => {
-    let manifest: Test;
+    let manifest: Manifest;
 
     beforeEach(async () => {
       await unlink(join(TEST_DIR, "/test2.t.js"));
@@ -93,16 +93,21 @@ describe('manifest-generator', () => {
     it('rewrites the manifest', () => {
       expect(manifest.children.length).toEqual(1)
       expect(manifest.children[0]).toEqual({ path: './tmp/manifest-generator/test1.t.js', description: 'hello' });
+      expect(manifest.errors).toHaveLength(0);
     });
   });
 
-  describe('no default export specs', () => {
-    let manifest: Manifest;
+  // describe('no default export specs', () => {
+  //   let manifest: Manifest;
 
-    beforeEach(async () => {
-      await writeFile(path.join(TEST_DIR , "/test4.t.js"), "module.exports = { namedExport: { description: 'test' } };");
-      await actions.receive(delegate, { event: "update" });
-      manifest = await loadManifest();
-    });
-  })
+  //   beforeEach(async () => {
+  //     await writeFile(path.join(TEST_DIR , "/test4.t.js"), "module.exports = { namedExport: { description: 'test' } };");
+  //     await actions.receive(delegate, { event: "update" });
+  //     manifest = await loadManifest();
+  //   });
+
+  //   it('adds errors to the manifest', () => {
+
+  //   })
+  // })
 });
